+++ src/lib/components/MobileNav.svelte
@@ -20,8 +20,8 @@
     },
     {
       href: '/tasks',
-      label: 'Tasks',
-      icon: 'M5 12l5 5L20 7',
+      label: 'Calendar',
+      icon: 'M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 0 0 2-2V7a2 2 0 0 0-2-2H5a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2z',
       match: (p) => p.startsWith('/tasks')
     },
     {